EventToken accepts event_index u32 explicitly, remove the expect

+++ crates/sui-indexer-alt-graphql/src/api/subscription.rs
@@ -193,7 +193,7 @@ impl Subscription {
                                 let cursor = EventToken::cursor(
                                     processed.summary.sequence_number,
                                     tx.tx_sequence_number,
-                                    idx as u64,
+                                    idx as u32,
                                 )
                                 .encode_cursor();
                                 yield Ok(Edge::new(
